refactor: move Storage
