+++ src/calendar/calendarController.ts
@@ -1,6 +1,7 @@
 import EventRecord from "./types/eventrecord";
 import Tag from "./types/tag";
 import Status from "./types/status";
+import Storage from "./storage";
 
 export interface ICalendarController {
   addEvent(event: EventRecord): Promise<number>;
@@ -8,11 +9,15 @@ export interface ICalendarController {
   getEventList(dateFrom: Date, dateTo: Date): Promise<EventRecord[] | null>;
   getEventList(tag: Tag): Promise<EventRecord[] | null>;
   getEventList(status: Status): Promise<EventRecord[] | null>;
+  getEventList(
+    dateFromOrTagOrStatus: Date | Tag | Status,
+    dateTo?: Date
+  ): Promise<EventRecord[] | null>;
   updateEvent(id: number, newEvent: EventRecord): Promise<EventRecord | null>;
   deleteEvent(id: number): Promise<number | null>;
 }
 
-export class CalendarController implements ICalendarStorage {
+export class CalendarController implements ICalendarController {
   private storage = new Storage();
 
   async addEvent(event: EventRecord): Promise<number> {
@@ -29,11 +34,13 @@ export class CalendarController implements ICalendarStorage {
 
   async getEventList(
     dateFromOrTagOrStatus: Date | Tag | Status,
-    dateTo: Date
+    dateTo?: Date
   ): Promise<EventRecord[] | null> {
     if (dateFromOrTagOrStatus instanceof Date) {
       const dateFrom = dateFromOrTagOrStatus;
-      return this.storage.getItemsByDate(dateFrom, dateTo);
+      if (dateTo instanceof Date) {
+        return this.storage.getItemsByDate(dateFrom, dateTo);
+      }
     } else if (dateFromOrTagOrStatus in Tag) {
       const tag = dateFromOrTagOrStatus as Tag;
       return this.storage.getItemsByTag(tag);
@@ -47,94 +54,5 @@ export class CalendarController implements ICalendarStorage {
     newEvent: EventRecord
   ): Promise<EventRecord | null> {
     return this.storage.updateItem(id, newEvent);
-  }
-}
-
-class Storage<T extends EventRecord = EventRecord> {
-  private idGenerator = numberGenerator();
-
-  add(item: T): number {
-    const nextId = this.idGenerator.next().value;
-    const newItem = { ...item, id: nextId };
-    localStorage.setItem(nextId, JSON.stringify(newItem));
-    return nextId;
-  }
-
-  getItem(id: number): T | null {
-    const item = localStorage.getItem(id.toString());
-    return item === null
-      ? null
-      : JSON.parse(item, (key: string, value: any) => {
-          if (key === "date") {
-            return new Date(value);
-          }
-          return value;
-        });
-  }
-
-  updateItem(id: number, item: T): T | null {
-    const idStr = id.toString();
-    if (idStr in localStorage) {
-      localStorage.setItem(idStr, JSON.stringify(item));
-      return item;
-    }
-    return null;
-  }
-
-  deleteItem(id: number): number | null {
-    const idStr = id.toString();
-    if (idStr in localStorage) {
-      localStorage.removeItem(idStr);
-      return id;
-    }
-    return null;
-  }
-
-  getItemsByDate(dateFrom: Date, dateTo: Date): T[] | null {
-    const keys = Object.keys(localStorage);
-    const result: T[] | null = [];
-    for (const key of keys) {
-      const item = this.getItem(Number(key)) as T;
-      const date = item.date;
-      if (date >= dateFrom && date <= dateTo) {
-        result.push(item);
-      }
-    }
-
-    return result.length > 0 ? result : null;
-  }
-
-  getItemsByTag(tag: Tag): T[] | null {
-    const keys = Object.keys(localStorage);
-    const result: T[] | null = [];
-    for (const key of keys) {
-      const item = this.getItem(Number(key)) as T;
-      const curTag = item.tag;
-      if (curTag === tag) {
-        result.push(item);
-      }
-    }
-
-    return result.length > 0 ? result : null;
-  }
-
-  getItemsByStatus(status: Status): T[] | null {
-    const keys = Object.keys(localStorage);
-    const result: T[] | null = [];
-    for (const key of keys) {
-      const item = this.getItem(Number(key)) as T;
-      const curStatus = item.status;
-      if (curStatus === status) {
-        result.push(item);
-      }
-    }
-
-    return result.length > 0 ? result : null;
-  }
-}
-
-function* numberGenerator(): Generator<number> {
-  for (let i = 1; ; i++) {
-    yield i;
   }
 }
